+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Upgraded build/test images to use Golang 1.13 instead of 1.11.
   [cyberark/summon-aws-secrets#27](https://github.com/cyberark/summon-aws-secrets/issues/27)
 
+### Fixed
+- Made installer script more robust to target environments.
+  [cyberark/summon-aws-secrets#16](https://github.com/cyberark/summon-aws-secrets/issues/16)
+
 ## [0.3.0] - 2019-03-06
 ### Added
 - Converted to go modules
@@ -27,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/cyberark/conjur-api-go/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/cyberark/conjur-api-go/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/cyberark/conjur-api-go/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/cyberark/conjur-api-go/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/cyberark/conjur-api-go/compare/v0.1.0...v0.2.0